Add all exercises fromex01 to ex08

+++ ex06/lib/typing.rb
@@ -1,5 +1,5 @@
 def typing(grade)
   puts "Digite sua nota: "
-  grade = gets.chomp.to_i
+  grade = gets.chomp.to_f
   grade
 end
